requirements.txt
chromadb==1.0.15
Flask==3.1.1
langchain==0.3.26
openai==1.95.1
python-dotenv==1.1.1
